Improved rendering (#33)
This includes lots of minor improvements, mainly to the hint system.
When removing a block, an 'X' hint will be shown. Hints are also drawn
on top of the terrain by default (configurable). When a warning or error
occurs, the hint where the problem occured will be tinted red or yellow.